# Improve Regex performance (in particular RegexOptions.Compiled)